Added plain html text format and CKEditor profile
An even simpler version than basic html to be used by users when adding
comments or reviews. It corresponds more or less to the plain text
format, but with the CKEditor toolbar for more easily adding bold and
italic text, or blockquote. A simple link plugin should be added as well.